refactor: extract force-save protocol into forcesave.js
Move MSG_FLUSH_REQUEST/RESPONSE constants, decodeFlushAck,
sendFlushRequest, waitForWsConnection, and forceSave out of
index.js into a dedicated forcesave.js module. index.js
re-exports nothing — it imports forceSave and attaches it
to daContent. Test file updated to import forceSave from
the new module path.
