+++ test/unit/blocks/edit/prose/index.test.js
@@ -5,8 +5,8 @@ import { setNx } from '../../../../../scripts/utils.js';
 import initProse, {
   createConnection,
   createAwarenessStatusWidget,
-  forceSave,
 } from '../../../../../blocks/edit/prose/index.js';
+import { forceSave } from '../../../../../blocks/edit/prose/forcesave.js';
 
 // initProse lazily imports da-library.js, which (a) builds URLs from
 // `${getNx()}/...` and (b) calls loadLibrary() at module import time.
